pyproject.toml
[project]
name = "padelstats"
version = "0.1.0"
description = "Padel stats static site generator"
authors = [
    {name = "Andreas Helbech Kleist"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "requests (>=2.33.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
